Extract Bookmark class to separate file and make independant of TextEditor
